refactor: remove unnecessary current
